fix: dont change composefile, just test for existance

+++ run_docker.sh
@@ -5,10 +5,10 @@ set -euxo pipefail
 # Required for otbr-docker
 sudo modprobe ip6table_filter
 
-# Adjust /dev/ttyACM* in compose file (a bit hacky, but oh well)
-dev_file=$(ls /dev/ttyACM* | head -n 1)
-sed -i -r -e "s#/dev/ttyACM.#${dev_file}#" docker-compose.yml
-
+if [ ! -f /dev/ttyACM0 ] ; then 
+	echo "File /dev/ttyACM0 does not exist"
+	exit 1
+fi
 
 docker container prune
